OT-5 auth logic

+++ internal/service/api/user/service.go
@@ -2,6 +2,7 @@ package user
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/go-openapi/strfmt"
@@ -9,7 +10,10 @@ import (
 
 	"otusgruz/internal/models"
 	query "otusgruz/internal/repo"
+	"otusgruz/pkg/http"
 )
+
+var ErrNoPermission = errors.New("No permission to perform action")
 
 type repo interface {
 	GetUser(ctx context.Context, guid uuid.UUID) (query.User, error)
@@ -36,6 +40,12 @@ func NewService(repo repo) Service {
 }
 
 func (s *service) GetUser(ctx context.Context, guid uuid.UUID) (*models.UserData, error) {
+	ctxUserGUID := http.UserGUIDFromContext(ctx)
+
+	if ctxUserGUID == uuid.Nil || ctxUserGUID != guid {
+		return nil, ErrNoPermission
+	}
+
 	res, err := s.repo.GetUser(ctx, guid)
 	if err != nil {
 		return nil, fmt.Errorf("getting user: %w", err)
@@ -50,6 +60,12 @@ func (s *service) GetUser(ctx context.Context, guid uuid.UUID) (*models.UserData
 }
 
 func (s *service) DeleteUser(ctx context.Context, guid uuid.UUID) (*models.DefaultStatusResponse, error) {
+	ctxUserGUID := http.UserGUIDFromContext(ctx)
+
+	if ctxUserGUID == uuid.Nil || ctxUserGUID != guid {
+		return nil, ErrNoPermission
+	}
+
 	err := s.repo.DeleteUser(ctx, guid)
 	if err != nil {
 		return nil, fmt.Errorf("deleting user: %w", err)
@@ -62,6 +78,12 @@ func (s *service) DeleteUser(ctx context.Context, guid uuid.UUID) (*models.Defau
 }
 
 func (s *service) UpdateUser(ctx context.Context, guid uuid.UUID, info *models.UserCreateParams) (*models.DefaultStatusResponse, error) {
+	ctxUserGUID := http.UserGUIDFromContext(ctx)
+
+	if ctxUserGUID == uuid.Nil || ctxUserGUID != guid {
+		return nil, ErrNoPermission
+	}
+
 	err := s.repo.UpdateUser(ctx, query.UpdateUserParams{
 		Guid:       guid,
 		Occupation: info.Occupation,
@@ -78,6 +100,12 @@ func (s *service) UpdateUser(ctx context.Context, guid uuid.UUID, info *models.U
 }
 
 func (s *service) CreateUser(ctx context.Context, info *models.UserCreateParams) (*models.DefaultStatusResponse, error) {
+	ctxUserGUID := http.UserGUIDFromContext(ctx)
+
+	if ctxUserGUID == uuid.Nil {
+		return nil, ErrNoPermission
+	}
+
 	err := s.repo.InsertUser(ctx, query.InsertUserParams{
 		Guid:       uuid.New(),
 		Occupation: info.Occupation,
